Added cartesian trajectory into move_pose_server. Debugged titration_vol uint32 issue.

+++ src/pipette_server.cpp
@@ -52,7 +52,7 @@ private:
         pub->publish(msg);
     }
 
-    inline void help_publish_uint32(rclcpp::Publisher<std_msgs::msg::UInt32>::SharedPtr pub, uint16_t value) {
+    inline void help_publish_uint32(rclcpp::Publisher<std_msgs::msg::UInt32>::SharedPtr pub, uint32_t value) {
         std_msgs::msg::UInt32 msg;
         msg.data = value;
         pub->publish(msg);
